Port the field model: wires, shapes, patterns and appearance
The "wires" half of the field system — one primitive taking a value from
another by id — plus the closure it needs: the remaining shapes, the vertex
patterns, and the appearance/fill/visibility/colour config.

Link resolution is the substance. LinkResolver.validate rejects forward
references, self-references, unknown targets and duplicate ids; the resolvers
then read radius, offset, scale, phase, orbit and appearance off the target.
Upstream checked this with a main() that printed ticks and crosses and lived
in src/main/java, so it never ran in a build. Its three cases are now the
first three tests in LinkModelTest, with ten more covering the resolver
surface it left alone — including that buildIndex(list, i) stops at i, which
is what makes a forward reference unresolvable rather than merely reported.

SimplePrimitive came across too. Without it the library would ship a
Primitive interface and a resolver with no way to make a Primitive.

Three classes were dropped rather than mangled: ColorTheme and
ColorThemeRegistry take a Minecraft Identifier, and ColorResolver reaches
into the mod's own config. FieldColor.mix was dropped as well, for a
different reason — its javadoc promised a 50/50 blend while it delegated to
Minecraft's ColorHelper.mix, and those are not the same operation. It has no
callers, Minecraft's source is not readable from here to settle which was
right, and lerp(0.5f, a, b) covers the documented intent. Guessing would have
been worse than leaving it out.

ColorSupport grew the six primitives FieldColor actually needed: white,
withAlpha, scaleRgb in both arities, and lerp. These are reimplementations
rather than moves, so they are tested — brightening saturates instead of
wrapping to black, and lerp rounds rather than truncates, both of which a
mask-based implementation gets silently wrong.

Two bugs in the port script, each of which had already bitten twice:

- The ColorHelper rewrite was a catch-all mapping every method to argb. That
  type-checks at the name but not the arity, so the errors surfaced far from
  the cause. It is now per-method.
- Imports were dropped after the package rewrite, which renamed the very
  imports the drop-list was looking for. Order reversed.

It also now strips private helpers taking a JsonObject, which are dead once
the public JSON methods are gone, and the empty for/if blocks left behind
where a Logging call was the only statement.

The javadoc is clean: 26 warnings cleared, most of them dangling @see links
to renderers and appliers that stay in the mod and will never be in the
library, which would have misled anyone reading the published API docs.

163 tests, no failures, no warnings.

+++ core/src/main/java/net/cyberpunk042/mcshaders/core/shape/ShapeMath.java
@@ -151,7 +151,7 @@ public final class ShapeMath {
      * @param theta Polar angle (0 = top pole, π = bottom pole)
      * @param phi Azimuthal angle (0 to 2π)
      * @param radius Base radius
-     * @param length Polar axis ratio (1 = sphere height, <1 = squashed, >1 = stretched)
+     * @param length Polar axis ratio (1 = sphere height, &lt;1 = squashed, &gt;1 = stretched)
      * @param bulge Equatorial bulge factor (0 = sphere width, 0.5 = 50% wider equator)
      */
     public static float[] spheroidWithBulge(float theta, float phi, float radius, 
@@ -613,7 +613,7 @@ public final class ShapeMath {
         
         for (int i = 0; i < count; i++) {
             // Fibonacci spiral distribution (even coverage)
-            float t = (float) (i + 0.5f) / (float) count;
+            float t = (i + 0.5f) / (float) count;
             float puffTheta = (float) Math.acos(1.0f - 2.0f * t);
             float puffPhi = goldenAngle * i;
             
@@ -695,7 +695,7 @@ public final class ShapeMath {
         float primaryRadius = 0.6f + smoothness * 0.4f;  // Wide puffs
         
         for (int i = 0; i < primaryCount; i++) {
-            float t = (float) (i + 0.5f) / (float) primaryCount;
+            float t = (i + 0.5f) / (float) primaryCount;
             float pTheta = (float) Math.acos(1.0f - 2.0f * t);
             float pPhi = goldenAngle * i;
             
@@ -723,7 +723,7 @@ public final class ShapeMath {
         float secondaryRadius = 0.3f + smoothness * 0.2f;
         
         for (int i = 0; i < secondaryCount; i++) {
-            float t = (float) (i + 0.5f) / (float) secondaryCount;
+            float t = (i + 0.5f) / (float) secondaryCount;
             float pTheta = (float) Math.acos(1.0f - 2.0f * t);
             float pPhi = goldenAngle * i * 1.618f;  // Different spiral
             
@@ -773,7 +773,7 @@ public final class ShapeMath {
         
         for (int i = 0; i < cellCount; i++) {
             // Cell center (Fibonacci spiral)
-            float t = (float) (i + 0.5f) / (float) cellCount;
+            float t = (i + 0.5f) / (float) cellCount;
             float cTheta = (float) Math.acos(1.0f - 2.0f * t);
             float cPhi = goldenAngle * i;
             
@@ -906,7 +906,7 @@ public final class ShapeMath {
             
             for (int i = 0; i < craterCount; i++) {
                 // Crater center position (Fibonacci spiral for even distribution)
-                float t = (float) (i + 0.5f) / (float) craterCount;
+                float t = (i + 0.5f) / (float) craterCount;
                 float craterTheta = (float) Math.acos(1.0f - 2.0f * t);
                 float craterPhi = goldenAngle * i;
